+++ BranchSDK/BNCSystemObserver.m
@@ -63,6 +63,7 @@
     __block NSString *token = nil;
     
 #if !TARGET_OS_TV
+#if !TARGET_OS_MACCATALYST
     if (@available(iOS 14.3, *)) {
 
         // We are getting reports on iOS 14.5 that this API can hang, adding a short timeout for now.
@@ -83,6 +84,7 @@
             BNCLogDebug([NSString stringWithFormat:@"AppleAttributionToken request timed out"]);
         }
     }
+#endif
 #endif
     
     return token;