+++ ChangeLog.md
@@ -1,5 +1,13 @@
 Branch iOS SDK Change Log
 
+v.2.1.1
+
+Branch iOS SDK 2.1.1 contains changes to how URL query parameters, like GBRAID and GCLID, are parsed and stored as well as a small fix for Mac Catalyst.
+
+- SDK-1800 - Extract GBRAID & GCLID from Branch referred uri/url
+    * Added new class to parse valid query parameters from URLs and append them to requests.
+- SDK-1747 - Added if available check for SKAN on Mac Catalyst
+
 v.2.1.0
 
 Branch iOS SDK 2.1.0 contains improvements to testing and plugin support (Unity, RN, etc). Most clients will see no change.